requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
idna==3.10
pycparser==2.22
requests==2.32.3
schwabdev==2.4.4
tzdata==2024.2
urllib3==2.2.3
websockets==14.1
python-dotenv==1.0.1
Flask==2.0.1
Werkzeug==2.0.1
pandas==2.2.3
numpy==2.1.3
matplotlib==3.10.0
plotly==5.24.1
yfinance==0.2.50